Split resolver replay kinds
